requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
gunicorn==20.1.0
psycopg2==2.9.5
alembic==1.8.1
python-dotenv==0.21.0